Split generated dashboards by component
This commit changes `exec` and `DashboardWriter` to use `DashboardResult` throughout and
use a `component` field to split generated dashboards into dirs.

This is useful in cases where you are only deploying dashboards for a
subset of the components offered in this repo.

Update README with the new dashboards + commands.

Tries to keep the end user API mostly the same.

Signed-off-by: Saswata Mukherjee <saswataminsta@yahoo.com>