Upgrade Perses go-sdk to v0.51, address breaking mods
In preparation for Perses v0.51.0, this commit upgrades the Perses
go-sdk module to v0.51.0-rc.0.

This is a breaking change, as all of the plugins that the dashboards
defined in this repo rely on, have been moved out of the main
perses/perses repo, and into https://github.com/perses/plugins.

This commit modifies the modules we rely on, and fixes that breaking
upgrade.

This upgrade is also needed, to allow us to add units to tables used in
these dashboards.

Signed-off-by: Saswata Mukherjee <saswataminsta@yahoo.com>