rules: Make API friendlier to downstream importers
This commit modifies some of the BuildXRules functions to be based off
of a NewXRulesBuilder method which returns the actual builder and
PrometheusRule object. This is ideal for downstream users who have their
own generators/writer.

This commit also adds several selector Options to different mixin rules,
for more downstream flexbility.

Finally also removes the generic type signaling, as that is
auto-detected.

Signed-off-by: Saswata Mukherjee <saswataminsta@yahoo.com>